requirements.txt
mlx>=0.0.1
numpy>=1.20.0,<2.0.0
rdkit-pypi>=2022.9.1
pandas>=1.5.0
scipy>=1.9.0
